feat: refactor Blog module to use BlogPost as aggregate root

+++ src/Modules/Blog/Explorer.Blog.Core/UseCases/BlogService.cs
@@ -26,7 +26,7 @@ namespace Explorer.Blog.Core.UseCases
             return _mapper.Map<BlogDto>(created);
         }
 
-        public BlogDto UpdateBlog(int id, CreateUpdateBlogDto dto, int userId)
+        public BlogDto UpdateBlog(long id, CreateUpdateBlogDto dto, int userId)
         {
             var blog = _repository.Get(id);
 
@@ -42,7 +42,7 @@ namespace Explorer.Blog.Core.UseCases
             return _mapper.Map<BlogDto>(blog);
         }
 
-        public BlogDto Get(int id)
+        public BlogDto Get(long id)
         {
             var blog = _repository.Get(id);
 
@@ -58,7 +58,7 @@ namespace Explorer.Blog.Core.UseCases
             return _mapper.Map<IEnumerable<BlogDto>>(blogs);
         }
 
-        public void DeleteBlog(int id, int userId)
+        public void DeleteBlog(long id, int userId)
         {
             var blog = _repository.Get(id);
 
@@ -71,7 +71,7 @@ namespace Explorer.Blog.Core.UseCases
             _repository.Delete(id);
         }
 
-        public void Publish(int id, int userId)
+        public void Publish(long id, int userId)
         {
             var blog = _repository.Get(id);
             if (blog == null)
@@ -84,7 +84,7 @@ namespace Explorer.Blog.Core.UseCases
             _repository.Update(blog);
         }
 
-        public void Archive(int id, int userId)
+        public void Archive(long id, int userId)
         {
             var blog = _repository.Get(id);
             if (blog == null)
